feat: fetch token timeseries data

+++ module-domain/src/main/kotlin/io/klaytn/finder/domain/mysql/set4/TokenTimeSeries.kt
@@ -9,7 +9,7 @@ import javax.persistence.Table
 @Table(name = "token_time_series")
 data class TokenTimeSeries(
     @Column()
-    val tokenInfoId: Int,
+    val tokenInfoId: Long,
 
     @Column()
     val symbol: String,
